[store] refactor: move netowk to network.rs
