+++ fusestore/store/src/meta_service/mod.rs
@@ -7,6 +7,7 @@ pub mod cmd;
 pub mod errors;
 pub mod log_entry;
 pub mod meta_service_impl;
+pub mod network;
 pub mod placement;
 pub mod raft_txid;
 pub mod raftmeta;
@@ -20,11 +21,11 @@ pub use errors::RetryableError;
 pub use errors::ShutdownError;
 pub use log_entry::LogEntry;
 pub use meta_service_impl::MetaServiceImpl;
+pub use network::Network;
 pub use placement::Placement;
 pub use raft_txid::RaftTxId;
 pub use raftmeta::MetaNode;
 pub use raftmeta::MetaStore;
-pub use raftmeta::Network;
 pub use snapshot::Snapshot;
 pub use state_machine::Node;
 pub use state_machine::Slot;
