+++ fusestore/store/src/meta_service/raftmeta.rs
@@ -9,16 +9,10 @@ use std::sync::Arc;
 
 use async_raft::async_trait::async_trait;
 use async_raft::config::Config;
-use async_raft::raft::AppendEntriesRequest;
-use async_raft::raft::AppendEntriesResponse;
 use async_raft::raft::ClientWriteRequest;
 use async_raft::raft::Entry;
 use async_raft::raft::EntryPayload;
-use async_raft::raft::InstallSnapshotRequest;
-use async_raft::raft::InstallSnapshotResponse;
 use async_raft::raft::MembershipConfig;
-use async_raft::raft::VoteRequest;
-use async_raft::raft::VoteResponse;
 use async_raft::storage::CurrentSnapshotData;
 use async_raft::storage::HardState;
 use async_raft::storage::InitialState;
@@ -28,7 +22,6 @@ use async_raft::LogId;
 use async_raft::NodeId;
 use async_raft::Raft;
 use async_raft::RaftMetrics;
-use async_raft::RaftNetwork;
 use async_raft::RaftStorage;
 use async_raft::SnapshotMeta;
 use common_exception::prelude::ErrorCode;
@@ -43,7 +36,6 @@ use common_runtime::tokio::sync::RwLockReadGuard;
 use common_runtime::tokio::sync::RwLockWriteGuard;
 use common_runtime::tokio::task::JoinHandle;
 use common_tracing::tracing;
-use tonic::transport::channel::Channel;
 
 use crate::meta_service::AppliedState;
 use crate::meta_service::Cmd;
@@ -51,8 +43,8 @@ use crate::meta_service::LogEntry;
 use crate::meta_service::MetaServiceClient;
 use crate::meta_service::MetaServiceImpl;
 use crate::meta_service::MetaServiceServer;
+use crate::meta_service::Network;
 use crate::meta_service::Node;
-use crate::meta_service::RaftMes;
 use crate::meta_service::RetryableError;
 use crate::meta_service::ShutdownError;
 use crate::meta_service::Snapshot;
@@ -61,59 +53,21 @@ use crate::meta_service::StateMachine;
 const ERR_INCONSISTENT_LOG: &str =
     "a query was received which was expecting data to be in place which does not exist in the log";
 
-impl tonic::IntoRequest<RaftMes> for AppendEntriesRequest<LogEntry> {
-    fn into_request(self) -> tonic::Request<RaftMes> {
-        let mes = RaftMes {
-            data: serde_json::to_string(&self).expect("fail to serialize"),
-            error: "".to_string(),
-        };
-        tonic::Request::new(mes)
-    }
-}
-impl tonic::IntoRequest<RaftMes> for InstallSnapshotRequest {
-    fn into_request(self) -> tonic::Request<RaftMes> {
-        let mes = RaftMes {
-            data: serde_json::to_string(&self).expect("fail to serialize"),
-            error: "".to_string(),
-        };
-        tonic::Request::new(mes)
-    }
-}
-impl tonic::IntoRequest<RaftMes> for VoteRequest {
-    fn into_request(self) -> tonic::Request<RaftMes> {
-        let mes = RaftMes {
-            data: serde_json::to_string(&self).expect("fail to serialize"),
-            error: "".to_string(),
-        };
-        tonic::Request::new(mes)
-    }
-}
-
-impl From<RetryableError> for RaftMes {
-    fn from(err: RetryableError) -> Self {
-        let error = serde_json::to_string(&err).expect("fail to serialize");
-        RaftMes {
-            data: "".to_string(),
-            error,
-        }
-    }
-}
-
 /// An in-memory storage system implementing the `async_raft::RaftStorage` trait.
 pub struct MetaStore {
     /// The ID of the Raft node for which this memory storage instances is configured.
-    id: NodeId,
+    pub id: NodeId,
     /// The Raft log.
-    log: RwLock<BTreeMap<u64, Entry<LogEntry>>>,
+    pub log: RwLock<BTreeMap<u64, Entry<LogEntry>>>,
     /// The Raft state machine.
-    state_machine: RwLock<StateMachine>,
+    pub state_machine: RwLock<StateMachine>,
     /// The current hard state.
-    hard_state: RwLock<Option<HardState>>,
+    pub hard_state: RwLock<Option<HardState>>,
 
-    snapshot_index: Arc<Mutex<u64>>,
+    pub snapshot_index: Arc<Mutex<u64>>,
 
     /// The current snapshot.
-    current_snapshot: RwLock<Option<Snapshot>>,
+    pub current_snapshot: RwLock<Option<Snapshot>>,
 }
 
 impl MetaStore {
@@ -454,84 +408,6 @@ impl RaftStorage<LogEntry, AppliedState> for MetaStore {
             }
             None => Ok(None),
         }
-    }
-}
-
-pub struct Network {
-    sto: Arc<MetaStore>,
-}
-
-impl Network {
-    pub fn new(sto: Arc<MetaStore>) -> Network {
-        Network { sto }
-    }
-
-    #[tracing::instrument(level = "info", skip(self), fields(myid=self.sto.id))]
-    pub async fn make_client(
-        &self,
-        node_id: &NodeId,
-    ) -> anyhow::Result<MetaServiceClient<Channel>> {
-        let addr = self.sto.get_node_addr(node_id).await?;
-        tracing::info!("connect: id={}: {}", node_id, addr);
-        let client = MetaServiceClient::connect(format!("http://{}", addr)).await?;
-        tracing::info!("connected: id={}: {}", node_id, addr);
-        Ok(client)
-    }
-}
-
-#[async_trait]
-impl RaftNetwork<LogEntry> for Network {
-    #[tracing::instrument(level = "info", skip(self), fields(myid=self.sto.id))]
-    async fn append_entries(
-        &self,
-        target: NodeId,
-        rpc: AppendEntriesRequest<LogEntry>,
-    ) -> anyhow::Result<AppendEntriesResponse> {
-        tracing::debug!("append_entries req to: id={}: {:?}", target, rpc);
-
-        let mut client = self.make_client(&target).await?;
-        let resp = client.append_entries(rpc).await;
-        tracing::debug!("append_entries resp from: id={}: {:?}", target, resp);
-
-        let resp = resp?;
-        let mes = resp.into_inner();
-        let resp = serde_json::from_str(&mes.data)?;
-
-        Ok(resp)
-    }
-
-    #[tracing::instrument(level = "info", skip(self), fields(myid=self.sto.id))]
-    async fn install_snapshot(
-        &self,
-        target: NodeId,
-        rpc: InstallSnapshotRequest,
-    ) -> anyhow::Result<InstallSnapshotResponse> {
-        tracing::debug!("install_snapshot req to: id={}", target);
-
-        let mut client = self.make_client(&target).await?;
-        let resp = client.install_snapshot(rpc).await;
-        tracing::debug!("install_snapshot resp from: id={}: {:?}", target, resp);
-
-        let resp = resp?;
-        let mes = resp.into_inner();
-        let resp = serde_json::from_str(&mes.data)?;
-
-        Ok(resp)
-    }
-
-    #[tracing::instrument(level = "info", skip(self), fields(myid=self.sto.id))]
-    async fn vote(&self, target: NodeId, rpc: VoteRequest) -> anyhow::Result<VoteResponse> {
-        tracing::debug!("vote req to: id={} {:?}", target, rpc);
-
-        let mut client = self.make_client(&target).await?;
-        let resp = client.vote(rpc).await;
-        tracing::info!("vote: resp from id={} {:?}", target, resp);
-
-        let resp = resp?;
-        let mes = resp.into_inner();
-        let resp = serde_json::from_str(&mes.data)?;
-
-        Ok(resp)
     }
 }
